split _bayesflow.py: _bayesflow_common.py + _bayesflow_posteriors.py
'_bayesflow' read as the generic module but actually held the NPE/posterior
machinery plus the shared helpers the likelihoods module imported -- a
misleading asymmetry against _bayesflow_likelihoods.py (and a bare rename
would have left the likelihoods module importing from 'posteriors'). Now:

- _bayesflow_common.py: the shared bridge -- lazy keras-pinned import,
  train-time validation, seeded-training RNG bracket, offline (theta, y)
  simulation, internal adapter keying. _ensure_bayesflow's error message
  generalized (it serves three entry points now).
- _bayesflow_posteriors.py (git mv, history preserved): AmortizedMethod,
  the NPE/FMPE/CMPE network factory, the NPE adapter, per-field bijectors,
  BayesFlowModel, learn_amortized_posterior.
- _bayesflow_likelihoods.py: unchanged content; imports from the common
  module (no sibling cross-dependency).

tests/inference/test_bayesflow.py renamed to test_bayesflow_posteriors.py
(tests mirror source, STYLE_GUIDE 8.1); ci bayesflow job and the private
test import updated. Pure reorganization: 67 tests (43 + 24) pass
unchanged; import probpipe stays lazy.
