+++ .github/workflows/ci.yml
@@ -351,13 +351,13 @@ jobs:
         run: uv sync --frozen --python ${{ matrix.python-version }} --extra dev --extra nutpie --extra bayesflow
 
       - name: Run BayesFlow backend tests
-        run: uv run pytest tests/inference/test_bayesflow.py tests/inference/test_bayesflow_likelihoods.py --cov-report=xml
+        run: uv run pytest tests/inference/test_bayesflow_posteriors.py tests/inference/test_bayesflow_likelihoods.py --cov-report=xml
 
       - name: Upload coverage reports to Codecov
         # The test job's upload measures the BayesFlow backend as uncovered: it
         # does not install [bayesflow], so these tests importorskip out there.
         # Upload this leg's report too — Codecov's per-commit union then counts
-        # _bayesflow.py toward patch coverage. 3.12 only, matching the test job.
+        # the _bayesflow_* modules toward patch coverage. 3.12 only, matching the test job.
         if: matrix.python-version == '3.12'
         uses: codecov/codecov-action@v5
         with:
