+++ probpipe/inference/__init__.py
@@ -21,7 +21,7 @@ from ._minibatch import MinibatchedDistribution
 # first call, so these eager imports stay cheap; the trained artifacts dispatch
 # via ``SupportsConditioning`` (NPE) or plug into ``SimpleModel`` as
 # ``Likelihood`` components (NLE/NRE) -- no inference-registry methods needed.
-from ._bayesflow import BayesFlowModel, learn_amortized_posterior
+from ._bayesflow_posteriors import BayesFlowModel, learn_amortized_posterior
 from ._bayesflow_likelihoods import (
     BayesFlowLikelihood, BayesFlowRatio,
     learn_amortized_likelihood, learn_amortized_ratio,
