+++ tests/inference/test_bayesflow_posteriors.py
